+++ app/src/main/java/com/example/myapplication/DetailsFragment.kt
@@ -1,0 +1,37 @@
+package com.example.myapplication
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+//import com.example.myapplication.databinding.ActivityDetailsBinding
+import com.example.myapplication.databinding.FragmentDetailsBinding
+import com.example.myapplication.databinding.FragmentHomeBinding
+
+class DetailsFragment : Fragment() {
+        private lateinit var binding: FragmentDetailsBinding
+        private lateinit var film: Film
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentDetailsBinding.inflate(inflater,container,false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setFilmsDetails()
+    }
+
+    private fun setFilmsDetails() {
+
+        film = arguments?.get("film") as Film
+        binding.detailsToolbar.title = film.title
+        binding.detailsPoster.setImageResource(film.poster)
+        binding.detailsDescription.text = film.description
+    }
+}
