+++ app/src/main/res/values/strings.xml
@@ -9,5 +9,7 @@
     <string name="menu_favorites_title"> Избранное </string>
     <string name="menu_watch_later_title"> Смотреть позже </string>
     <string name="menu_selections_title"> Рекомендации </string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
