+++ app/src/main/res/values/themes.xml
@@ -1,8 +1,7 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
     <style name="Base.Theme.MyApplication" parent="Theme.MaterialComponents.Light.DarkActionBar">
-        <!-- Customize your light theme here. -->
-        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <!--style name="Base.Theme.MyApplication" parent="Theme.MaterialComponents.Light.DarkActionBar"-->
     </style>
 
     <style name="Theme.MyApplication" parent="Base.Theme.MyApplication" />
